Plutarch.Api

Description

Note

The PValue and PMap-related functionality can be found in other modules, as these clash with the Plutarch prelude. These should be imported qualified.

Contexts

newtype PScriptContext (s :: S)

Since: 2.0.0

Constructors

PScriptContext (Term s (PDataRecord '["txInfo" ':= PTxInfo, "purpose" ':= PScriptPurpose]))

newtype PTxInfo (s :: S)

Constructors

PTxInfo (Term s (PDataRecord '["inputs" ':= PBuiltinList PTxInInfo, "referenceInputs" ':= PBuiltinList PTxInInfo, "outputs" ':= PBuiltinList PTxOut, "fee" ':= PValue 'Sorted 'Positive, "mint" ':= PValue 'Sorted 'NoGuarantees, "dcert" ':= PBuiltinList PDCert, "wdrl" ':= PMap 'Unsorted PStakingCredential PInteger, "validRange" ':= PInterval PPosixTime, "signatories" ':= PBuiltinList (PAsData PPubKeyHash), "redeemers" ':= PMap 'Unsorted PScriptPurpose PRedeemer, "datums" ':= PMap 'Unsorted PDatumHash PDatum, "id" ':= PTxId]))

data PScriptPurpose (s :: S)

Since: 2.0.0

Constructors

PMinting (Term s (PDataRecord '["_0" ':= PCurrencySymbol]))

PSpending (Term s (PDataRecord '["_0" ':= PTxOutRef]))

PRewarding (Term s (PDataRecord '["_0" ':= PStakingCredential]))

PCertifying (Term s (PDataRecord '["_0" ':= PDCert]))

Tx

newtype PTxOutRef (s :: S)

Reference to a transaction output, with an index referencing which exact output we mean.

Since: 2.0.0

Constructors

PTxOutRef (Term s (PDataRecord '["id" ':= PTxId, "idx" ':= PInteger]))

newtype PTxOut (s :: S)

Since: 2.0.0

Constructors

PTxOut (Term s (PDataRecord '["address" ':= PAddress, "value" ':= PValue 'Sorted 'Positive, "datum" ':= POutputDatum, "referenceScript" ':= PMaybeData PScriptHash]))

newtype PTxId (s :: S)

Since: 2.0.0

Constructors

PTxId (Term s (PDataRecord '["_0" ':= PByteString]))

newtype PTxInInfo (s :: S)

An input of the pending transaction.

Since: 2.0.0

Constructors

PTxInInfo (Term s (PDataRecord '["outRef" ':= PTxOutRef, "resolved" ':= PTxOut]))

data POutputDatum (s :: S)

Since: 2.0.0

Constructors

PNoOutputDatum (Term s (PDataRecord '[]))

POutputDatumHash (Term s (PDataRecord '["datumHash" ':= PDatumHash]))

POutputDatum (Term s (PDataRecord '["outputDatum" ':= PDatum]))
Inline datum as per CIP-0032

Script

Types

newtype PDatum (s :: S)

Since: 2.0.0

Constructors

PDatum (Term s PData)

newtype PDatumHash (s :: S)

Since: 2.0.0

Constructors

PDatumHash (Term s PByteString)

newtype PRedeemer (s :: S)

Since: 2.0.0

Constructors

PRedeemer (Term s PData)

newtype PRedeemerHash (s :: S)

Since: 2.0.0

Constructors

PRedeemerHash (Term s PByteString)

newtype PScriptHash (s :: S)

Since: 2.0.0

Constructors

PScriptHash (Term s PByteString)

Functions

scriptHash :: Script -> ScriptHash

Hash a script, appending the Plutus V2 prefix.

Since: 2.0.0

datumHash :: Datum -> DatumHash

Since: 2.0.0

redeemerHash :: Redeemer -> RedeemerHash

Since: 2.0.0

dataHash :: forall (a :: Type). ToData a => a -> BuiltinByteString

Since: 2.0.0

Value

newtype PValue (keys :: KeyGuarantees) (amounts :: AmountGuarantees) (s :: S)

Since: 2.0.0

Constructors

PValue (Term s (PMap keys PCurrencySymbol (PMap keys PTokenName PInteger)))

data AmountGuarantees

Since: 2.0.0

Constructors

NoGuarantees

NonZero

Positive

newtype PCurrencySymbol (s :: S)

Since: 2.0.0

Constructors

PCurrencySymbol (Term s PByteString)

newtype PTokenName (s :: S)

Since: 2.0.0

Constructors

PTokenName (Term s PByteString)

DCert

data PDCert (s :: S)

Since: 2.0.0

Constructors

PDCertDelegRegKey (Term s (PDataRecord '["_0" ':= PStakingCredential]))

PDCertDelegDeRegKey (Term s (PDataRecord '["_0" ':= PStakingCredential]))

PDCertDelegDelegate (Term s (PDataRecord '["_0" ':= PStakingCredential, "_1" ':= PPubKeyHash]))

PDCertPoolRegister (Term s (PDataRecord '["_0" ':= PPubKeyHash, "_1" ':= PPubKeyHash]))

PDCertPoolRetire (Term s (PDataRecord '["_0" ':= PPubKeyHash, "_1" ':= PInteger]))

PDCertGenesis (Term s (PDataRecord '[]))

PDCertMir (Term s (PDataRecord '[]))

Assoc map

Types

newtype PMap (keysort :: KeyGuarantees) (k :: PType) (v :: PType) (s :: S)

Since: 2.0.0

Constructors

PMap (Term s (PBuiltinList (PBuiltinPair (PAsData k) (PAsData v))))

data KeyGuarantees

Since: 2.0.0

Constructors

Sorted

Unsorted

data Commutativity

Since: 2.0.0

Constructors

Commutative

NonCommutative

Address

data PCredential (s :: S)

Since: 2.0.0

Constructors

PPubKeyCredential (Term s (PDataRecord '["_0" ':= PPubKeyHash]))

PScriptCredential (Term s (PDataRecord '["_0" ':= PScriptHash]))

data PStakingCredential (s :: S)

Since: 2.0.0

Constructors

PStakingHash (Term s (PDataRecord '["_0" ':= PCredential]))

PStakingPtr (Term s (PDataRecord '["_0" ':= PInteger, "_1" ':= PInteger, "_2" ':= PInteger]))

newtype PAddress (s :: S)

Since: 2.0.0

Constructors

PAddress (Term s (PDataRecord '["credential" ':= PCredential, "stakingCredential" ':= PMaybeData PStakingCredential]))

Time

newtype PPosixTime (s :: S)

Since: 2.0.0

Constructors

PPosixTime (Term s PInteger)

Interval

newtype PInterval (a :: S -> Type) (s :: S)

Since: 2.0.0

Constructors

PInterval (Term s (PDataRecord '["from" ':= PLowerBound a, "to" ':= PUpperBound a]))

newtype PLowerBound (a :: S -> Type) (s :: S)

Since: 2.0.0

Constructors

PLowerBound (Term s (PDataRecord '["_0" ':= PExtended a, "_1" ':= PBool]))

newtype PUpperBound (a :: S -> Type) (s :: S)

Since: 2.0.0

Constructors

PUpperBound (Term s (PDataRecord '["_0" ':= PExtended a, "_1" ':= PBool]))

data PExtended (a :: S -> Type) (s :: S)

Since: 2.0.0

Constructors

PNegInf (Term s (PDataRecord '[]))

PFinite (Term s (PDataRecord '["_0" ':= a]))

PPosInf (Term s (PDataRecord '[]))

Crypto

Types

newtype PubKey

Since: 2.0.0

Constructors

PubKey

Fields

getPubKey :: LedgerBytes

Since: 2.0.0

newtype PPubKeyHash (s :: S)

Since: 2.0.0

Constructors

PPubKeyHash (Term s PByteString)

pubKeyHash :: PubKey -> PubKeyHash

Since: 2.0.0

Utilities

Types

data PMaybeData (a :: S -> Type) (s :: S)

Since: 2.0.0

Constructors

PDJust (Term s (PDataRecord '["_0" ':= a]))

PDNothing (Term s (PDataRecord '[]))

Functions

ptuple :: forall (a :: S -> Type) (b :: S -> Type) (s :: S). Term s (PAsData a :--> (PAsData b :--> PDataSum '['["_0" ':= a, "_1" ':= b]]))

Since: 2.0.0

Orphan instances

PConstantDecl Address

Instance details

Associated Types

type PConstantRepr Address

type PConstanted Address :: PType

Methods

pconstantToRepr :: Address -> PConstantRepr Address

pconstantFromRepr :: PConstantRepr Address -> Maybe Address

PConstantDecl ScriptPurpose

Instance details

Associated Types

type PConstantRepr ScriptPurpose

type PConstanted ScriptPurpose :: PType

Methods

pconstantToRepr :: ScriptPurpose -> PConstantRepr ScriptPurpose

pconstantFromRepr :: PConstantRepr ScriptPurpose -> Maybe ScriptPurpose

PConstantDecl Credential

Instance details

Associated Types

type PConstantRepr Credential

type PConstanted Credential :: PType

Methods

pconstantToRepr :: Credential -> PConstantRepr Credential

pconstantFromRepr :: PConstantRepr Credential -> Maybe Credential

PConstantDecl StakingCredential

Instance details

Associated Types

type PConstantRepr StakingCredential

type PConstanted StakingCredential :: PType

Methods

pconstantToRepr :: StakingCredential -> PConstantRepr StakingCredential

pconstantFromRepr :: PConstantRepr StakingCredential -> Maybe StakingCredential

PConstantDecl PubKeyHash

Instance details

Associated Types

type PConstantRepr PubKeyHash

type PConstanted PubKeyHash :: PType

Methods

pconstantToRepr :: PubKeyHash -> PConstantRepr PubKeyHash

pconstantFromRepr :: PConstantRepr PubKeyHash -> Maybe PubKeyHash

PConstantDecl DCert

Instance details

Associated Types

type PConstantRepr DCert

type PConstanted DCert :: PType

Methods

pconstantToRepr :: DCert -> PConstantRepr DCert

pconstantFromRepr :: PConstantRepr DCert -> Maybe DCert

PConstantDecl Datum

Instance details

Associated Types

type PConstantRepr Datum

type PConstanted Datum :: PType

Methods

pconstantToRepr :: Datum -> PConstantRepr Datum

pconstantFromRepr :: PConstantRepr Datum -> Maybe Datum

PConstantDecl DatumHash

Instance details

Associated Types

type PConstantRepr DatumHash

type PConstanted DatumHash :: PType

Methods

pconstantToRepr :: DatumHash -> PConstantRepr DatumHash

pconstantFromRepr :: PConstantRepr DatumHash -> Maybe DatumHash

PConstantDecl Redeemer

Instance details

Associated Types

type PConstantRepr Redeemer

type PConstanted Redeemer :: PType

Methods

pconstantToRepr :: Redeemer -> PConstantRepr Redeemer

pconstantFromRepr :: PConstantRepr Redeemer -> Maybe Redeemer

PConstantDecl ScriptHash

Instance details

Associated Types

type PConstantRepr ScriptHash

type PConstanted ScriptHash :: PType

Methods

pconstantToRepr :: ScriptHash -> PConstantRepr ScriptHash

pconstantFromRepr :: PConstantRepr ScriptHash -> Maybe ScriptHash

PConstantDecl POSIXTime

Instance details

Associated Types

type PConstantRepr POSIXTime

type PConstanted POSIXTime :: PType

Methods

pconstantToRepr :: POSIXTime -> PConstantRepr POSIXTime

pconstantFromRepr :: PConstantRepr POSIXTime -> Maybe POSIXTime

PConstantDecl TxId

Instance details

Associated Types

type PConstantRepr TxId

type PConstanted TxId :: PType

Methods

pconstantToRepr :: TxId -> PConstantRepr TxId

pconstantFromRepr :: PConstantRepr TxId -> Maybe TxId

PConstantDecl TxOutRef

Instance details

Associated Types

type PConstantRepr TxOutRef

type PConstanted TxOutRef :: PType

Methods

pconstantToRepr :: TxOutRef -> PConstantRepr TxOutRef

pconstantFromRepr :: PConstantRepr TxOutRef -> Maybe TxOutRef

PConstantDecl ScriptContext

Instance details

Associated Types

type PConstantRepr ScriptContext

type PConstanted ScriptContext :: PType

Methods

pconstantToRepr :: ScriptContext -> PConstantRepr ScriptContext

pconstantFromRepr :: PConstantRepr ScriptContext -> Maybe ScriptContext

PConstantDecl TxInInfo

Instance details

Associated Types

type PConstantRepr TxInInfo

type PConstanted TxInInfo :: PType

Methods

pconstantToRepr :: TxInInfo -> PConstantRepr TxInInfo

pconstantFromRepr :: PConstantRepr TxInInfo -> Maybe TxInInfo

PConstantDecl TxInfo

Instance details

Associated Types

type PConstantRepr TxInfo

type PConstanted TxInfo :: PType

Methods

pconstantToRepr :: TxInfo -> PConstantRepr TxInfo

pconstantFromRepr :: PConstantRepr TxInfo -> Maybe TxInfo

PConstantDecl OutputDatum

Instance details

Associated Types

type PConstantRepr OutputDatum

type PConstanted OutputDatum :: PType

Methods

pconstantToRepr :: OutputDatum -> PConstantRepr OutputDatum

pconstantFromRepr :: PConstantRepr OutputDatum -> Maybe OutputDatum

PConstantDecl TxOut

Instance details

Associated Types

type PConstantRepr TxOut

type PConstanted TxOut :: PType

Methods

pconstantToRepr :: TxOut -> PConstantRepr TxOut

pconstantFromRepr :: PConstantRepr TxOut -> Maybe TxOut

PConstantData a => PConstantDecl (Interval a)

Instance details

Associated Types

type PConstantRepr (Interval a)

type PConstanted (Interval a) :: PType

Methods

pconstantToRepr :: Interval a -> PConstantRepr (Interval a)

pconstantFromRepr :: PConstantRepr (Interval a) -> Maybe (Interval a)

PConstantData a => PConstantDecl (LowerBound a)

Instance details

Associated Types

type PConstantRepr (LowerBound a)

type PConstanted (LowerBound a) :: PType

Methods

pconstantToRepr :: LowerBound a -> PConstantRepr (LowerBound a)

pconstantFromRepr :: PConstantRepr (LowerBound a) -> Maybe (LowerBound a)

PConstantData a => PConstantDecl (UpperBound a)

Instance details

Associated Types

type PConstantRepr (UpperBound a)

type PConstanted (UpperBound a) :: PType

Methods

pconstantToRepr :: UpperBound a -> PConstantRepr (UpperBound a)

pconstantFromRepr :: PConstantRepr (UpperBound a) -> Maybe (UpperBound a)

PConstantData a => PConstantDecl (Maybe a)

Instance details

Associated Types

type PConstantRepr (Maybe a)

type PConstanted (Maybe a) :: PType

Methods

pconstantToRepr :: Maybe a -> PConstantRepr (Maybe a)

pconstantFromRepr :: PConstantRepr (Maybe a) -> Maybe (Maybe a)