Plutarch.Api.AssocMap

Description

This module is designed to be imported qualified, as many of its identifiers clash with the Plutarch prelude.

Types

newtype PMap (keysort :: KeyGuarantees) (k :: PType) (v :: PType) (s :: S)

Since: 2.0.0

Constructors

PMap (Term s (PBuiltinList (PBuiltinPair (PAsData k) (PAsData v))))

data KeyGuarantees

Since: 2.0.0

Constructors

Sorted

Unsorted

data Commutativity

Since: 2.0.0

Constructors

Commutative

NonCommutative

Functions

Creation

pempty :: Term s (PMap 'Sorted k v)

Construct an empty PMap.

Since: 2.0.0

Transformation

passertSorted :: forall (k :: S -> Type) (v :: S -> Type) (any :: KeyGuarantees) (s :: S). (POrd k, PIsData k, PIsData v) => Term s (PMap any k v :--> PMap 'Sorted k v)

Given a PMap of uncertain order, yield a PMap that is known to be sorted.

Since: 2.0.0

pmap :: forall (g :: KeyGuarantees) (k :: S -> Type) (a :: S -> Type) (b :: S -> Type) (s :: S). (PIsData a, PIsData b) => Term s ((a :--> b) :--> (PMap g k a :--> PMap g k b))

Applies a function to every value in the map, much like map.

Since: 2.0.0

pmapData :: forall (g :: KeyGuarantees) (k :: S -> Type) (a :: S -> Type) (b :: S -> Type) (s :: S). Term s ((PAsData a :--> PAsData b) :--> (PMap g k a :--> PMap g k b))

As pmap, but over Data representations.

Since: 2.0.0

pmapMaybe :: forall (g :: KeyGuarantees) (k :: S -> Type) (a :: S -> Type) (b :: S -> Type) (s :: S). (PIsData a, PIsData b) => Term s ((a :--> PMaybe b) :--> (PMap g k a :--> PMap g k b))

Maps and filters the map, much like mapMaybe.

Since: 2.0.0

pmapMaybeData :: forall (g :: KeyGuarantees) (k :: S -> Type) (a :: S -> Type) (b :: S -> Type) (s :: S). Term s ((PAsData a :--> PMaybe (PAsData b)) :--> (PMap g k a :--> PMap g k b))

As pmapMaybe, but over Data representation.

Since: 2.0.0

Relational lift

pcheckBinRel :: forall (k :: S -> Type) (v :: S -> Type) (s :: S). (POrd k, PIsData k, PIsData v) => Term s ((v :--> (v :--> PBool)) :--> (v :--> (PMap 'Sorted k v :--> (PMap 'Sorted k v :--> PBool))))

Given a comparison function and a "zero" value, check whether a binary relation holds over 2 sorted PMaps.

Important note

This is primarily intended to be used with PValue. We assume that the comparison behaves like a comparison would (thus, being at least a partial order, or possibly a total order or equivalence), and that the starting value does not break it. Use with extreme care.

Since: 2.0.0

Folds

pall :: forall (any :: KeyGuarantees) (k :: S -> Type) (v :: S -> Type) (s :: S). PIsData v => Term s ((v :--> PBool) :--> (PMap any k v :--> PBool))

Verifies all values in the map satisfy the given predicate.

Since: 2.0.0

Combination

punionResolvingCollisionsWith :: forall (k :: S -> Type) (v :: S -> Type) (s :: S). (POrd k, PIsData k, PIsData v) => Commutativity -> Term s ((v :--> (v :--> v)) :--> (PMap 'Sorted k v :--> (PMap 'Sorted k v :--> PMap 'Sorted k v)))

Build the union of two PMaps, merging values that share the same key using the given function.

Since: 2.0.0

Queries

pnull :: forall (any :: KeyGuarantees) (k :: S -> Type) (v :: S -> Type) (s :: S). Term s (PMap any k v :--> PBool)

Tests whether the map is empty.

Since: 2.0.0

Orphan instances

(PConstantData k, PConstantData v, Ord k) => PConstantDecl (Map k v)
Since: 2.0.0

Instance details

Associated Types

type PConstantRepr (Map k v)

type PConstanted (Map k v) :: PType

Methods

pconstantToRepr :: Map k v -> PConstantRepr (Map k v)

pconstantFromRepr :: PConstantRepr (Map k v) -> Maybe (Map k v)