Plutarch.Api.Utils

Description

Useful tools that aren't part of the Plutarch API per se, but get used in multiple places.

Documentation

newtype Mret (a :: S -> Type) (s :: S)

Term, but with its type arguments flipped. This is a useful helper for defining PTryFrom instances.

For example, consider the PTryFrom instance for PTokenName:

instance PTryFrom PData (PAsData PTokenName) where
   type PTryFromExcess PData (PAsData PTokenName) = Mret PTokenName

We need to do this because PTryFromExcess expects something of kind S -> Type, but Term has kind S -> (S -> Type) -> Type, which doesn't quite fit. By using Mret, we end up with something of kind (S -> Type) -> S -> Type, which fits.

The name is just Term written backwards.

Since: 2.0.0

Constructors

Mret (Term s a)