Plutarch.Api.Value

Description

Value-related functionality. In order to keep the interface efficient and safe at the same time, there is a type-level distinction between PValues that are guaranteed to be properly normalized and those that provide no such guarantee.

Also for efficiency reasons, the Ada-specific functions assume that there can be only one token name for the Ada currency symbol, and they don't check whether it matches adaToken.

Types

newtype PValue (keys :: KeyGuarantees) (amounts :: AmountGuarantees) (s :: S)

Since: 2.0.0

Constructors

PValue (Term s (PMap keys PCurrencySymbol (PMap keys PTokenName PInteger)))

newtype PCurrencySymbol (s :: S)

Since: 2.0.0

Constructors

PCurrencySymbol (Term s PByteString)

newtype PTokenName (s :: S)

Since: 2.0.0

Constructors

PTokenName (Term s PByteString)

data AmountGuarantees

Since: 2.0.0

Constructors

NoGuarantees

NonZero

Positive

Functions

Transformation

passertPositive :: forall (kg :: KeyGuarantees) (ag :: AmountGuarantees) (s :: S). Term s (PValue kg ag :--> PValue kg 'Positive)

Given a PValue, either construct another PValue with the same contents and a proof that all amounts in it are positive, or error.

Since: 2.0.0

pforgetPositive :: forall (a :: AmountGuarantees) (k :: KeyGuarantees) (s :: S). Term s (PValue k 'Positive) -> Term s (PValue k a)

'Forget' that a Value has an only-positive guarantee.

Since: 2.0.0

pnormalize :: forall (any :: AmountGuarantees) (s :: S). Term s (PValue 'Sorted any :--> PValue 'Sorted 'NonZero)

Normalize the argument to contain no zero quantity nor empty token map.

Since: 2.0.0

Partial ordering

pcheckBinRel :: forall (any0 :: AmountGuarantees) (any1 :: AmountGuarantees) (s :: S). Term s ((PInteger :--> (PInteger :--> PBool)) :--> (PValue 'Sorted any0 :--> (PValue 'Sorted any1 :--> PBool)))

Given a description of a relation on amounts, check whether that relation holds over sorted PValues.

Important note

This is intended for use with boolean comparison functions, which must define at least a partial order (total orders and equivalences are acceptable as well). Use of this with anything else is not guaranteed to give anything resembling a sensible answer. Use with extreme care.

Since: 2.0.0

Combination

punionResolvingCollisionsWith :: forall (any0 :: AmountGuarantees) (any1 :: AmountGuarantees) (s :: S). Commutativity -> Term s ((PInteger :--> (PInteger :--> PInteger)) :--> (PValue 'Sorted any0 :--> (PValue 'Sorted any1 :--> PValue 'Sorted 'NoGuarantees)))

Combine two PValues applying the given function to any pair of quantities with the same asset class. Note that the result is _not_ normalized and may contain zero quantities.

Since: 2.0.0

Orphan instances

PConstantDecl CurrencySymbol

Instance details

Associated Types

type PConstantRepr CurrencySymbol

type PConstanted CurrencySymbol :: PType

Methods

pconstantToRepr :: CurrencySymbol -> PConstantRepr CurrencySymbol

pconstantFromRepr :: PConstantRepr CurrencySymbol -> Maybe CurrencySymbol

PConstantDecl TokenName

Instance details

Associated Types

type PConstantRepr TokenName

type PConstanted TokenName :: PType

Methods

pconstantToRepr :: TokenName -> PConstantRepr TokenName

pconstantFromRepr :: PConstantRepr TokenName -> Maybe TokenName

PConstantDecl Value

Instance details

Associated Types

type PConstantRepr Value

type PConstanted Value :: PType

Methods

pconstantToRepr :: Value -> PConstantRepr Value

pconstantFromRepr :: PConstantRepr Value -> Maybe Value